requirements.txt
pyserial==3.5
numpy==1.19.4
matplotlib==3.4.1
siriuspy
